Fix parse_failure method in workflows.rb
If workflows.rb encounters a problem parsing a parameter -- e.g., if
an element is nil -- it will call the parse_failure method. That
method tries to print out the values of the "wfparam" and "parameter"
Ruby variables. But it fails because those variables are not passed in
to that method (and are out of scope). So Test Kitchen exits with this
cryptic error:

  Failed to complete #create action: [undefined local variable or
  method `wfparam' for VcoWorkflows::Workflow:Class]

This PR simply passes those variables in to the parse_failure method
so it can display those values.